Move NameType logic to api object
